+++ lib/views/auth_gate.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:only_job/views/home/home.dart';
+import 'package:only_job/views/home/temphome.dart';
 import 'package:only_job/views/authenticate/authenticate.dart';
 import 'package:provider/provider.dart';
 import 'package:only_job/models/user.dart';
@@ -15,7 +15,7 @@ class _AuthGateState extends State<AuthGate> {
   @override
   Widget build(BuildContext context) {
     final user = Provider.of<User?>(context);
-    if(user != null){
+    if (user != null) {
       return Home();
     } else {
       return Authenticate();
