+++ lib/views/authenticate/client_or_employee.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class ClientOrEmployee extends StatelessWidget {
+  ClientOrEmployee({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+    );
+  }
+}
