+++ lib/views/authenticate/register_or_login.dart
@@ -1,0 +1,44 @@
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+import 'package:only_job/views/authenticate/client_or_employee.dart';
+import 'package:only_job/views/authenticate/login.dart';
+
+class RegisterLogin extends StatelessWidget {
+  const RegisterLogin({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            // Image(image: image),
+            // Image(image: image),
+            ElevatedButton(
+              onPressed: () {
+                Navigator.of(context).push(
+                  MaterialPageRoute(
+                    builder: (context) => ClientOrEmployee(),
+                  ),
+                );
+              },
+              child: Text('Create New Account'),
+            ),
+            ElevatedButton(
+              onPressed: () {
+                Navigator.of(context).push(
+                  MaterialPageRoute(
+                    builder: (context) => Login(),
+                  ),
+                );
+              },
+              child: Text('Log in'),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
