--- lib/views/register_login.dart
@@ -1,23 +1,0 @@
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
-
-class RegisterLogin extends StatelessWidget {
-  const RegisterLogin({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            // Image(image: image),
-            // Image(image: image),
-          ],
-        
-        ),
-      ),
-    );
-  }
-}
